requirements.txt
colorama>=0.4.6
termcolor>=2.3.0
pyfiglet>=0.8.post1
requests>=2.31.0
python-whois>=0.8.0
dnspython>=2.4.2
python-nmap>=0.7.1
beautifulsoup4>=4.12.2
lxml>=4.9.3
pysocks>=1.7.1
geoip2>=4.7.0
argparse>=1.4.0
python-dateutil>=2.8.2
tqdm>=4.66.1
tabulate>=0.9.0
validators>=0.22.0
psutil>=5.9.5
urllib3>=2.0.4